add jupyter lab 3.6.5 to tests
Tests take now a JUPYTER_TYPE that determines the if executed with
jupyter lab or or notebook. By that the class names are determined.
Previous tests can be now execued with the tox environment
tests-notebook-7, the jupyter lab tests with tests-lab-3 tests. Tests
have been extended where they needed to be to also work with jupyter
lab.