Tests: Add rerun functionality for tests
We add pytest-rerun to tests and allow the tests to be rerun up to 3
times until they pass. The tests for lab 4 are less stable and fail at
different points for different reasons that are hard to fix. By allowing
the tests to pass with a rerun we avoid these kind of issues.