Core: Fix code is None in CodeExercise with ExerciseRegistry
When the code is None in the CodeExercise and a ExerciseRegistry is
given, then an error is raised when the save cue box is created, because
the cue_widget is None. Since the cue widgets are set later on, we use a
dummy widget on initialization with is later replaced.