Fix non-clearing output when nothing is outputted (#60)
When the the update does not print anything the output is not cleared,
since it waits for an output to be cleared. We fix this by printing an
empty string to invoke the clearning of the output

* Ignore B040 for checks failing

Through an update linter complained that no error message is raised, but
we do not want to raise an error message when one of the checks raises
one.