Check results are now collected before handling
When multiple checks were handled only the output of the last check
remained. This is because all checks were individually handled and each
handling cleared the output. We now collect the check result including
the exceptions and then send it for handling.

We also now only print a summary of the results in the CheckRegistry as
the output was too extensive.