Fix for nested errors are not correctly resolved (#62)
When an exception is thrown within a WCI, the last frame is used as
context to add the input lines. This does not work if the exception is
thrown within a function within WCI. Therefore we now iterate through
all traceback frames to find the one corresponding to WCI.

The corresponding PR in WCI https://github.com/osscar-org/widget-code-input/pull/26
to solve it there.