Mirror update button description in public API of `CodeExercise`
To improve the API from a user perspective we use the same naming for
public functions that update a `CodeExercise` instantce as the buttons
as their description. This means renaming `run_update` to `run_code` and
duplicating the functionality of the `run_update` with the function
`update`, since the update button is called "Update" when no `code` is
given on initalization. Since `run_code` was already taken we rename it
tc `call_code`