Print empty char after updated code exercise (#80)
To enforce `clear_output` even when the update is not printing anything
we used `print("", end"")` this seems not to always work so we print an
empty char now. In the tests this introduces another output that makes
the tests fail checking the number of outputs. Because the number of
outputs is testing anything meaningful we remove these tests. There are
tests that check the joined texts of the outputs that cover the outputs
already good enough.