Fix figure not properly closed in CodeExercise with parameters (#94)
When a CodeExercise with parameters is created and a figure is created the
`<Figure size 640x480 with 0 Axes>` is displayed. This is prevented by closing
the figure when `CueFigure` is created. I am not sure why this happens in this
case. It could be related to interact. It can be fixed by moving the CueFigure
creation before the ParameterPanel creation.