Do not display the set `Set all references` button of the `CheckRegistry` by default (#44)
This button is only meant for teachers so it should be able to disable
it. For that we add the property `display_set_all_references_button` that
can be given on initilization or setted after. We also change the order
of the buttons so the `Check all widgets` button does not change its
position between enabling and disabling this option.