Update the check notebook
Remove old usage of `get_function_object`.

Remove section that refers how to compute output for a check because it
can be now just done with `code_ex.code()`.

Remove widget that raises error since it is confusing what is meant. The
connection with the check registry is not made.