Update the check notebook (#113)
Replace old usage of `get_function_object()(...)` with `code(...)`

Remove section that refers how to compute output for a check because it can be
now just done with `code_ex.code()`.

Remove widget that raises error since it is confusing what is meant. The
connection with the check registry is not made.