Disable run button if no update function and parameters are given
The run button was still displayed in this case but could only work
in case the code does not require any arguments. To enable the update
button one needs now to at least provide parameters (empty in case when
no parameters are needed). This enables the teacher to create exercises
that are only checkable.