Disable run button if no update function and parameters are given (#115)
The run button was still displayed in this case but could only work
when the code does not require any arguments. Now to enable the update
button one needs to at least provide parameters (empty in case when
no parameters are needed). This enables the teacher to create exercises
that are only checkable.